major modification and improvement of the Plot package to enable generic plot editability

+++ src/main/java/org/baratinage/ui/plot/IPlotItemRendererSettings.java
@@ -12,7 +12,7 @@ public interface IPlotItemRendererSettings {
 
     public float[] getLineDashArray();
 
-    public double getShapeSize();
+    public int getShapeSize();
 
     public ShapeType getShapeType();
